geoportal/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
geoalchemy2==0.14.2
shapely==2.0.2
geopandas==0.14.1 